Completed base website
Completed the base website which can connect to a database via MongoDB. Following this I have made a controller and model schema for the database with some limitations on what can be added to the database. I have also created the ability to add, remove and view all users in the database with ease. I have also created a simple CSS stylesheet and used EJS to create my website.